Sync integration tests (#3690)
* Upgraded RealmAppImport to use the realm-cli

* Changed to symlinked consumation for the electron app

* Upgraded mocha-remote and fixed integration tests

* Migrated to importing from "realm"

* Major refactoring to the integration test harnesses

* Simplified the react-native integration test app

* Made the import app utility return an app instance

* Adding a failing anonymous credentials test

* Attempt at running integration tests on CI

* Switched to lerna when installing environments and tests

* Updating some NPM scripts

* Fixing root package-lock.json

* Removed unneeded complexity in RN test app.

* Ignoring scripts when installing locally

* Upgraded typescript and package-locks

* Removed the regular dependency on realm and tests again

* Revert "Removed unneeded complexity in RN test app."

This reverts commit 16e04eb4287b82007d8e2cac81ab294a3b85163e.

* Fixed running ios tests via a --prefix

* Updated the README

* Run linting on sub-packages explicitly

* Scoping bootstrapping in GH workflow

* Propagating context as environments

* Adding skipIf to suites

* Running local only integration tests on CI

* Using lerna bootstrap instead of npm install

* Renamed localOnly to missingServer

* Moved integration tests to GHA

* Running iOS integration tests

* Caching xcframework + installing tests package

* Inhibiting warnings

* Using the mocha-github-actions-reporter

* Run in chrome-debugging mode

* Upgrading mocha remote

* Switch mode from context

* Updated script seperator to :

* No need to disconnect on end with autoDisconnect

* Caching XCode derived data

* Caching CocoaPods

* Include the CONTEXT environment variable in NPM script

* Try running just the Chrome Debugging

* Adding cache on Podfile.lcok

* Installing watchman

* Updated .lock pattern

* Adding start.js to pass arguments to mocha

* Debugging metro cache invalidation

* Adding an option to run chrome debugger headless

* Printing module verboseNames

* Logging stack on error

* Try / catching more.

* Cache the ios build folder

* Iterating on printing error stacks

* Removed caching ios build artifacts again

* Logging a stack when a message is logged

* Caching node_modules

* Try messing less with the RN cache

* Retrying cache of the ios build folder

* Debugging the tests

* Adding a tmate session for debugging

* Trying to catch error loading "realm"

* Debugging requiring "realm"

* Reworded the install script override

* Removed peer dependencies on "realm" and the tests

* Bumping the ios-build cache version

* No need for a mocha peer dependency

* Re-introducing the tests

* Propagate the error through

* Adding build:watch script to tests

* Try running tests on Android emulator

* Building and running tests for android

* Run on Electron as well

* Reexporting ANDROID_NDK_HOME

* Using xvfb-run

* Simplified electron harness

* Trying Android test again

* Trying a node matrix

* Update pr-integration-tests.yml

* Adding windows to the Node.js builds

* Adding steps to setup windows build env

* Use prebuilds for node and run in release and debug mode

* Adding runtime and target to package.json's config

* Upgraded mocha-remote-cli

* Retry opening the Metro Chrome debugging UI

* Run React Native in both release & debug

* Fixed RN Android job name

* Renamed flow file and name

* Upgrading mocha-remote

* Upgrading CocoaPods and Flipper-Folly

* Calling logcat before exiting on Android

* Updated README.md files

* Adding a Realm.Set when chrome debugging (#3704)

* Enable streaming logcat while testing

* Clarified some documentation

* Minor clean ups

* Xcode updated some project files

* Attempting to apply a fix to breaking ios builds

* Simplified skipIf

* Moved the local_prebuilds env

* Reusing a variable

* Disabled chrome debugging tests for now

* Update CHANGELOG.md

* Apply suggestions from code review

+++ packages/realm-app-importer/src/AppImporter.ts
@@ -20,22 +20,31 @@ import cp from "child_process";
 import fetch from "node-fetch";
 import path from "path";
 import fs from "fs-extra";
+import glob from "glob";
+import deepmerge from "deepmerge";
+
+/**
+ * First level keys are file globs and the values are objects that are spread over the content of the files matching the glob.
+ * @example { "config.json": { name: "overridden-name" }, "services/local-mongodb/rules/*.json": { database: "another-database" } }
+ */
+export type TemplateReplacements = Record<string, Record<string, unknown>>;
 
 /* eslint-disable no-console */
 
-export interface RealmAppImporterOptions {
+export interface AppImporterOptions {
     baseUrl: string;
     username: string;
     password: string;
-    stitchConfigPath: string;
+    realmConfigPath: string;
     appsDirectoryPath: string;
+    cleanUp?: boolean;
 }
 
-export class RealmAppImporter {
+export class AppImporter {
     private readonly baseUrl: string;
     private readonly username: string;
     private readonly password: string;
-    private readonly stitchConfigPath: string;
+    private readonly realmConfigPath: string;
     private readonly appsDirectoryPath: string;
 
     private accessToken: string | undefined;
@@ -44,34 +53,43 @@ export class RealmAppImporter {
         baseUrl,
         username,
         password,
-        stitchConfigPath,
+        realmConfigPath,
         appsDirectoryPath,
-    }: RealmAppImporterOptions) {
+        cleanUp = true,
+    }: AppImporterOptions) {
         this.baseUrl = baseUrl;
         this.username = username;
         this.password = password;
-        this.stitchConfigPath = stitchConfigPath;
+        this.realmConfigPath = realmConfigPath;
         this.appsDirectoryPath = appsDirectoryPath;
 
-        process.on("exit", () => {
-            // Remove any stitch configuration
-            if (fs.existsSync(this.stitchConfigPath)) {
-                console.log(`Deleting ${this.stitchConfigPath}`);
-                fs.removeSync(this.stitchConfigPath);
-            }
-            // If there is nothing the the apps directory, lets delete it
-            if (fs.existsSync(this.appsDirectoryPath)) {
-                console.log(`Deleting ${this.appsDirectoryPath}`);
-                fs.removeSync(this.appsDirectoryPath);
-            }
-        });
+        if (cleanUp) {
+            process.on("exit", () => {
+                // Remove any stitch configuration
+                if (fs.existsSync(this.realmConfigPath)) {
+                    console.log(`Deleting ${this.realmConfigPath}`);
+                    fs.removeSync(this.realmConfigPath);
+                }
+                // If there is nothing the the apps directory, lets delete it
+                if (fs.existsSync(this.appsDirectoryPath)) {
+                    console.log(`Deleting ${this.appsDirectoryPath}`);
+                    fs.removeSync(this.appsDirectoryPath);
+                }
+            });
+        }
     }
 
+    /**
+     * @param appTemplatePath The path to a template directory containing the configuration files needed to import the app.
+     * @param replacements An object with file globs as keys and a replacement object as values. Allows for just-in-time replacements of configuration parameters.
+     * @returns A promise of an object containing the app id.
+     */
     public async importApp(
         appTemplatePath: string,
+        replacements: TemplateReplacements = {},
     ): Promise<{ appId: string }> {
-        const { name: appName } = this.loadStichJson(appTemplatePath);
-        await this.login();
+        const { name: appName } = this.loadAppConfigJson(appTemplatePath);
+        await this.logIn();
         const groupId = await this.getGroupId();
 
         // Get or create an application
@@ -94,12 +112,14 @@ export class RealmAppImporter {
         const appPath = path.resolve(this.appsDirectoryPath, appId);
         // Copy over the app template
         this.copyAppTemplate(appPath, appTemplatePath);
+        // Apply any replacements to the files before importing from them
+        this.applyReplacements(appPath, replacements);
 
         // Import
-        this.stitchCli(
+        this.realmCli(
             "import",
             "--config-path",
-            this.stitchConfigPath,
+            this.realmConfigPath,
             "--base-url",
             this.baseUrl,
             "--app-name",
@@ -119,6 +139,10 @@ export class RealmAppImporter {
         return { appId };
     }
 
+    private get apiUrl() {
+        return `${this.baseUrl}/api/admin/v3.0`;
+    }
+
     private loadJson(filePath: string) {
         try {
             const content = fs.readFileSync(filePath, "utf8");
@@ -130,15 +154,15 @@ export class RealmAppImporter {
         }
     }
 
-    private loadStichJson(appTemplatePath: string) {
-        const stitchJsonPath = path.resolve(appTemplatePath, "stitch.json");
-        return this.loadJson(stitchJsonPath);
+    private loadAppConfigJson(appTemplatePath: string) {
+        const configJsonPath = path.resolve(appTemplatePath, "config.json");
+        return this.loadJson(configJsonPath);
     }
 
     private loadSecretsJson(appTemplatePath: string) {
-        const stitchJsonPath = path.resolve(appTemplatePath, "secrets.json");
-        if (fs.existsSync(stitchJsonPath)) {
-            return this.loadJson(stitchJsonPath);
+        const secretsJsonPath = path.resolve(appTemplatePath, "secrets.json");
+        if (fs.existsSync(secretsJsonPath)) {
+            return this.loadJson(secretsJsonPath);
         } else {
             return {};
         }
@@ -154,20 +178,35 @@ export class RealmAppImporter {
         }
     }
 
-    private stitchCli(...args: string[]) {
-        cp.spawnSync("stitch-cli", args, { stdio: "inherit" });
+    private applyReplacements(
+        appPath: string,
+        replacements: TemplateReplacements,
+    ) {
+        for (const [fileGlob, replacement] of Object.entries(replacements)) {
+            console.log(`Applying replacements to ${fileGlob}`);
+            const files = glob.sync(fileGlob, { cwd: appPath });
+            for (const relativeFilePath of files) {
+                const filePath = path.resolve(appPath, relativeFilePath);
+                const content = fs.readJSONSync(filePath);
+                const mergedContent = deepmerge(content, replacement);
+                fs.writeJSONSync(filePath, mergedContent, { spaces: 2 });
+            }
+        }
     }
 
-    private async login() {
-        const url = `${this.baseUrl}/api/admin/v3.0/auth/providers/local-userpass/login`;
-        const body = JSON.stringify({
-            username: this.username,
-            password: this.password,
-        });
+    private realmCli(...args: string[]) {
+        cp.spawnSync("realm-cli", args, { stdio: "inherit" });
+    }
+
+    private async logIn() {
+        const url = `${this.apiUrl}/auth/providers/local-userpass/login`;
         const response = await fetch(url, {
             method: "POST",
             headers: { "content-type": "application/json" },
-            body,
+            body: JSON.stringify({
+                username: this.username,
+                password: this.password,
+            }),
         });
         // Store the access and refresh tokens
         const responseBody = await response.json();
@@ -186,16 +225,12 @@ export class RealmAppImporter {
         refreshToken: string,
         accessToken: string,
     ) {
-        const stitchConfig = [
+        const realmConfig = [
             `public_api_key: ${username}`,
             `refresh_token: ${refreshToken}`,
             `access_token: ${accessToken}`,
         ];
-        fs.writeFileSync(
-            this.stitchConfigPath,
-            stitchConfig.join("\n"),
-            "utf8",
-        );
+        fs.writeFileSync(this.realmConfigPath, realmConfig.join("\n"), "utf8");
     }
 
     private async getProfile() {
